[wip]remove extra animals from the map
I don't see that having a bunch of extra animals taking up space on the
map adds to the fun of game play. And I find it annoying that units will
spend time attacking and killing them when there are other enemy units
nearby

* Should this change be made to the other farms?
* Suggestions as to a different approach, or different values that I've
used here?

Don't merge. This is wip, and up for discussion.