romans:guard tower:add ability to move short distances
Experimental change... this unit is given to romans at the start.
sometimes it's in the way of a resource and people feel it must be
destroyed immediately.

This change gives 5 EP to the tower, and costs 1 EP to move. It can be
moved a very short distance. Unfortunately, it applies to all guard
towers, not just the startup tower.