Romans xml clean (#183)
* romans:clean up xml

The AI romans build just fine without all this extra stuff in romans.xml

When the colosseum produces grace, the AI is far more likely to build
colosseums instead of eagle pillars, so removed grace production from
colosseum.

Moved the location of the "hold position" command. Some bug was
preventing that from being displayed when the mechanic was selected.

* added eagle pillar to starting units (so they'd have at least one resource-producing unit at start, like other factions)

I think this will be a little cleaner. The eagle pillar isn't of any use
offensively until later in the game anyway.